camptail: add verbose flag
Right now this just shows which rooms are entered. Which can
be handy if your config doesn't actually show backlog, and
you want to make sure campfire actually has you in the right
rooms. We should perhaps have a "rooms to always be in"
config, and it should enter them, too.