camptail: add --state option
This option will keep the room-state in a file from run to
run, so that you can periodically get new messages without
having to keep a streaming connection open.

Note that this is not connected to streaming at all. In
particular, streamed data will not go into the state file.
This is mostly out of laziness on my part, since it is not
clear when the state file should be written while streaming,
and for my purposes, this is about periodic polling, not
streaming.