Room: add transcript command
The recent command is quite limited; it will only pull the
last 100 messages. This is way too few for something like my
campmail-style grepping. What I really want is to say "this
is the last id I got, and show me everything in this room
since then".

Sadly, the transcript API doesn't seem to take a "since"
parameter, so we're stuck pulling the whole daily transcript
and grepping it ourselves.