use transcript to handle large (or multi-day) tails
Tailing via recent will get us at most the 100 last
messages. This is not enough for polling even every hour in
a busy room. Instead, let's use the transcript API to just
pull whole days.

We can start tracking the last day pulled in the RoomState
file, so we won't pull old days over and over again. But we
still have to keep pulling the _entire_ current day as it
gets added to, because there is no way to limit it at the
API level.