camptail: notice room changes during grep
Because grep treats the messages as a single stream,
regardless of room, if we are grepping with -A/-B, we may
end up showing unrelated messages from another room.

This has always been the case, and in the context of
matching "recent" messages, it is not that bad.
Conversations from another room might actually be relevant.
But with transcripts, they are almost certainly not, because
we lose all time locality.

And especially when you are grepping, it gives you a bunch
of irrelevant context lines to read through, which is
annoying. So let's just drop it.