pad "--all" days by one on either side
The Campfire API leaves unspecified the time zone in which a
transcript's day occurs. So after midnight local time, the
client may think it is now a new day, whereas the server
does not; we will miss messages sent during that time
period, because we cut-off our search by day. Padding
backwards 1 day solves that, at the expense of requesting
too much data.

Similarly, if it is not yet midnight on the client, the
server may already think it is "tomorrow", and querying by
day will not get new messages until the day counter rolls
over locally. This can be fixed by padding one day into the
future.