Campfire: don't feed raw scalar reference to curl
Newer versions of WWW::Curl are happy for us to hand a
scalar reference to WRITEDATA, and store the output in the
scalar. Older versions (including the one on Debian
squeeze) want a real filehandle.

It's not clear to me if this is related to the version of
curl, WWW::Curl, or perl. In any case, it's easy to work
around by asking perl to create a filehandle that outputs to
the scalar.