Campfire::Base: use scalar context for accessors
Older versions of XML::Smart were happy to access scalar XML
items in an array context. But since upgrading to jessie,
with XML::Smart 1.78, this now causes a perl error.

I'm not sure if this is a bug in XML::Smart, or a problem in
my code, but we can easily work around it by using a scalar
context (this means that the _accessor() helper can't be
used for arrays anymore, but I don't think any call sites do
so).